issue #1, #2, #3, #4 수정
