+++ app/posts/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 # Register your models here.
-from posts.models import SalesForm, ControlPoint, PostAddress, SecuritySafetyFacilities, PostRoom
+from posts.models import SalesForm, ControlPoint, PostAddress, SecuritySafetyFacilities, PostRoom, OptionItem
 
 
 class PostRoomAdmin(admin.ModelAdmin):
@@ -24,8 +24,13 @@ class SecuritySafetyFacilitiesAdmin(admin.ModelAdmin):
     list_display = ['pk', 'name']
 
 
+class OptionItemAdmin(admin.ModelAdmin):
+    list_display = ['pk', 'name', ]
+
+
 admin.site.register(PostRoom, PostRoomAdmin)
 admin.site.register(SalesForm, SalesFormAdmin)
 admin.site.register(ControlPoint, ControlPointAdmin)
 admin.site.register(PostAddress, PostAddressAdmin)
 admin.site.register(SecuritySafetyFacilities, SecuritySafetyFacilitiesAdmin)
+admin.site.register(OptionItem, OptionItemAdmin)
